Extract getextractCommand into its own file
